+++ src/services/content-org/learning-paths.ts
@@ -7,9 +7,14 @@ import { fetchByRailContentId, fetchMethodV2Structure } from '../sanity.js'
 import { addContextToContent } from '../contentAggregator.js'
 import {
   contentStatusCompleted,
+  contentsStatusCompleted,
   contentStatusReset,
+  getAllCompletedByIds,
   getProgressState,
 } from '../contentProgress.js'
+import { COLLECTION_TYPE, STATE } from "../sync/models/ContentProgress";
+import { SyncWriteDTO } from "../sync";
+import { ContentProgress } from "../sync/models";
 
 const BASE_PATH: string = `/api/content-org`
 const LEARNING_PATHS_PATH = `${BASE_PATH}/v1/user/learning-paths`
@@ -20,6 +25,20 @@ interface ActiveLearningPathResponse {
   active_learning_path_id: number,
 }
 
+interface DailySessionResponse {
+  user_id: number,
+  brand: string,
+  user_date: string
+  daily_session: DailySession[],
+  active_learning_path_id: number,
+  active_learning_path_created_at: string,
+}
+
+interface DailySession {
+  content_ids: number[],
+  learning_path_id: number,
+}
+
 /**
  * Gets today's daily session for the user.
  * @param brand
@@ -28,7 +47,7 @@ interface ActiveLearningPathResponse {
 export async function getDailySession(brand: string, userDate: Date) {
   const stringDate = userDate.toISOString().split('T')[0]
   const url: string = `${LEARNING_PATHS_PATH}/daily-session/get?brand=${brand}&userDate=${stringDate}`
-  return await fetchHandler(url, 'GET', null, null)
+  return await fetchHandler(url, 'GET', null, null) as DailySessionResponse
 }
 
 /**
@@ -45,7 +64,7 @@ export async function updateDailySession(
   const stringDate = userDate.toISOString().split('T')[0]
   const url: string = `${LEARNING_PATHS_PATH}/daily-session/create`
   const body = { brand: brand, userDate: stringDate, keepFirstLearningPath: keepFirstLearningPath }
-  return await fetchHandler(url, 'POST', null, body)
+  return await fetchHandler(url, 'POST', null, body) as DailySessionResponse
 }
 
 /**
@@ -54,7 +73,7 @@ export async function updateDailySession(
  */
 export async function getActivePath(brand: string) {
   const url: string = `${LEARNING_PATHS_PATH}/active-path/get?brand=${brand}`
-  return await fetchHandler(url, 'GET', null, null)
+  return await fetchHandler(url, 'GET', null, null) as ActiveLearningPathResponse
 }
 
 /**
@@ -65,7 +84,7 @@ export async function getActivePath(brand: string) {
 export async function startLearningPath(brand: string, learningPathId: number) {
   const url: string = `${LEARNING_PATHS_PATH}/active-path/set`
   const body = { brand: brand, learning_path_id: learningPathId }
-  return await fetchHandler(url, 'POST', null, body)
+  return await fetchHandler(url, 'POST', null, body) as ActiveLearningPathResponse
 }
 
 /**
@@ -85,9 +104,9 @@ export async function getEnrichedLearningPath(learningPathId) {
   const response = (await addContextToContent(
     fetchByRailContentId,
     learningPathId,
-    'learning-path-v2',
+    COLLECTION_TYPE.LEARNING_PATH,
     {
-      collection: { id: learningPathId, type: 'learning-path-v2' },
+      collection: { id: learningPathId, type: COLLECTION_TYPE.LEARNING_PATH },
       dataField: 'children',
       dataField_includeParent: true,
       addProgressStatus: true,
@@ -220,8 +239,21 @@ export async function fetchLearningPathLessons(
   }
 }
 
+/**
+ * For an array of contentIds, fetch any content progress with state=completed,
+ * including other learning paths and a la carte progress.
+ *
+ * @param {number[]} contentIds The array of content IDs within the learning path
+ * @returns {Promise<number[]>} Array with completed content IDs
+ */
+export async function fetchLearningPathProgressCheckLessons(contentIds: number[]): Promise<number[]> {
+  let query = await getAllCompletedByIds(contentIds)
+  let completedProgress = query.data.map(progress => progress.content_id)
+  return contentIds.filter(contentId => completedProgress.includes(contentId))
+}
+
 interface completeMethodIntroVideo {
-  intro_video_response: Object | null,
+  intro_video_response: SyncWriteDTO<ContentProgress, any> | null,
   active_path_response: ActiveLearningPathResponse
 }
 /**
@@ -238,7 +270,7 @@ export async function completeMethodIntroVideo(introVideoId: number, brand: stri
   response.intro_video_response = await completeIfNotCompleted(introVideoId)
 
   const methodStructure = await fetchMethodV2Structure(brand)
-  const learningPathId = methodStructure.learningPaths[0].id
+  const learningPathId = methodStructure.learning_paths[0].id
 
   response.active_path_response = await startLearningPath(brand, learningPathId)
 
@@ -247,9 +279,9 @@ export async function completeMethodIntroVideo(introVideoId: number, brand: stri
 }
 
 interface completeLearningPathIntroVideo {
-  intro_video_response: Object | null,
-  learning_path_reset_response: void | null,
-  lesson_import_response: Object[] | null
+  intro_video_response: SyncWriteDTO<ContentProgress, any> | null,
+  learning_path_reset_response: SyncWriteDTO<ContentProgress, any> | null,
+  lesson_import_response: SyncWriteDTO<ContentProgress, any> | null
 }
 /**
  * Handles completion of learning path intro video and other related actions.
@@ -266,23 +298,20 @@ export async function completeLearningPathIntroVideo(introVideoId: number, learn
 
   response.intro_video_response = await completeIfNotCompleted(introVideoId)
 
-  const collection = { id: learningPathId, type: 'learning-path-v2' }
+  const collection = { id: learningPathId, type: COLLECTION_TYPE.LEARNING_PATH }
 
   if (!lessonsToImport) {
-    // returns nothing now, but it will when watermelon comes 'round
     response.learning_path_reset_response = await contentStatusReset(learningPathId, collection)
 
   } else {
-    for (const contentId of lessonsToImport) {
-      response.lesson_import_response[contentId] = await contentStatusCompleted(contentId, collection)
-    }
+    response.lesson_import_response = await contentsStatusCompleted(lessonsToImport, collection)
   }
 
   return response
 }
 
 
-async function completeIfNotCompleted(contentId: number): Promise<Object | null> {
+async function completeIfNotCompleted(contentId: number): Promise<SyncWriteDTO<ContentProgress, any> | null> {
   const introVideoStatus = await getProgressState(contentId)
 
   return introVideoStatus !== 'completed' ? await contentStatusCompleted(contentId) : null